No such thing as "IAM Account" only "IAM Alias"
An AWS account can have an IAM Alias. But there is no such thing as an "IAM Account" in AWS. Hence, the title of this article should use "IAM Alias" instead, and be renamed as Managing IAM Aliases. 

If you use "IAM Account", then people will wrongly think that there is a different account for the IAM service.